# server/swagger/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Cascarita APIs",
    "version": "1.0.0",
    "description": "Backend Cascartia APIs"
  },
  "paths": {
...
    "/api/teams/groups/{group_id}": {
      "get": {
        "operationId": "getTeamsByGroupId",
        "summary": "Get teams By group ID",
        "description": "Get teams by group ID.",
        "parameters": [
          {
            "in": "path",
            "name": "group_id",
            "required": true,
            "schema": {
              "title": "group_id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TeamSchema"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Group not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessageSchema"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessageSchema"
                }
              }
            }
          }
        },
        "tags": ["Teams"]
      }
    },
    "/api/teams": {
      "post": {
        "operationId": "createTeam",
        "summary": "Create Team",
        "description": "Create a team.",
        "parameters": [],
        "responses": {
          "201": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TeamSchema"
                }
              }
            }
          },
          "400": {
            "description": "Team name already exists",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessageSchema"
                }
              }
            }
          },
          "404": {
            "description": "Group not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessageSchema"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessageSchema"
                }
              }
            }
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTeamSchema"
              }
            }
          },
          "required": true
        },
        "tags": ["Teams"]
      }
    },
    "/api/teams/{team_id}": {
      "get": {
        "operationId": "getTeam",
        "summary": "Get Team",
        "description": "Get a team.",
        "parameters": [
          {
            "in": "path",
            "name": "team_id",
            "required": true,
            "schema": {
              "title": "team_id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TeamSchema"
                }
              }
            }
          },
          "404": {
            "description": "Team not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessageSchema"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessageSchema"
                }
              }
            }
          }
        },
        "tags": ["Teams"]
      },
      "put": {
        "operationId": "updateTeam",
        "summary": "Update Team",
        "description": "Update a team.",
        "parameters": [
          {
            "in": "path",
            "name": "team_id",
            "required": true,
            "schema": {
              "title": "team_id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TeamSchema"
                }
              }
            }
          },
          "404": {
            "description": "Team not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessageSchema"
                }
              }
            }
          },
          "400": {
            "description": "Team with name already exists",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessageSchema"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessageSchema"
                }
              }
            }
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTeamSchema"
              }
            }
          },
          "required": true
        },
        "tags": ["Teams"]
      },
      "delete": {
        "operationId": "deleteTeam",
        "summary": "Delete Team",
        "description": "Delete a team.",
        "parameters": [
          {
            "in": "path",
            "name": "team_id",
            "required": true,
            "schema": {
              "title": "team_id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "OK"
          },
          "404": {
            "description": "Team not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessageSchema"
                }
              }
            }
          },
          "400": {
            "description": "Failure to delete team",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessageSchema"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessageSchema"
                }
              }
            }
          }
        },
        "tags": ["Teams"]
      }
    },
    "/api/teams/seasons/{season_id}/divisions/{division_id}": {
      "get": {
        "operationId": "getTeamsBySeasonDivisionId",
        "summary": "Get Teams by Seasons and Divisions",
        "description": "Get seasons divisions by season ID and division ID.",
        "parameters": [
          {
            "in": "path",
            "name": "season_id",
            "required": true,
            "schema": {
              "title": "season_id",
              "type": "integer"
            }
          },
          {
            "in": "path",
            "name": "division_id",
            "required": true,
            "schema": {
              "title": "division_id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TeamSchema"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessageSchema"
                }
              }
            }
          }
        },
        "tags": ["Teams"]
      }
    },
...
  },
  "components": {
    "schemas": {
      "UpdateTeamSchema": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "nullable": true
          },
          "group_id": {
            "type": "integer"
          },
          "team_logo": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "CreateTeamSchema": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "nullable": true
          },
          "group_id": {
            "type": "integer"
          },
          "team_logo": {
            "type": "string",
            "nullable": true
          }
        },
        "required": ["group_id", "name", "team_logo"],
        "tags": ["Teams"]
      },
      "TeamSchema": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "group_id": {
            "type": "integer"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          },
          "team_logo": {
            "type": "string",
            "nullable": true
          }
        },
        "tags": ["Teams"]
      },
...
      "ErrorMessageSchema": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          }
        },
        "tags": ["Utility"]
      },
...
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "servers": null
}